DataNetworkModels.jl

Documentation for DataNetworkModels.jl

DataNetworkModels.channels_load — Method

channels_load(speeds_flow_step, channels_to_flow_steps)

return load for each channel channels_to_flow_steps maps channel to flows and index of channel in flow load equals sum of all speeds

DataNetworkModels.final_hop_matrix — Function

final_hop_matrix(g, [hop_matrix])

return final hop-distance matrix for graph g

DataNetworkModels.find_flow_path — Method

find_flow_path(source, target, shortest_paths_to_neighbors)

return vector of nodes including source and target nodes guaranteed to have minimum length

DataNetworkModels.find_next_step — Method

find_next_step(source, target, shortest_paths_to_neighbors)

return random neighbor lying on the shortest path

DataNetworkModels.map_channels_to_flow_steps — Method

map_channels_to_flow_steps(flow_steps_to_channel, channel_number)

return vector of tuples: flow indices and step indices for each channel

DataNetworkModels.map_shortest_paths_to_neighbors — Method

map_shortest_paths_to_neighbors(g, hop_matrix)

return vector{vector{vector{Int}}} - [source][target][neighbor] map each pair of (source, target) to subset of source neighbors who are part of one of the shortest path from source to target

DataNetworkModels.propagate_speeds! — Method

propagate_speeds!(speeds_flow_step, speeds_start)

update speeds for every flow and channel u[flow][start] := speeds_start[flow] u[flow][channel] := u[flow][channel - 1]

DataNetworkModels.regulate_speed! — Method

regulate_speed!(speeds_flow_step, current_channels_load, flow_steps_to_channel, max_channels_load)

regulate speed for each flow and channel so channelloads ≤ maxchannels_load

DataNetworkModels.update_hop_matrix! — Method

update_hop_matrix!(hop_matrix, g)

distributed (local) update of hop matrix

after L updates since last graph change hop matrix is guaranteed to provide true hop distances where L - maximum length of shortest paths

DataNetworkModels.update_speed! — Method

update_speed!(
    speeds_flow_step,
    speeds_start,
    channels_to_flow_steps,
    flow_steps_to_channel,
    max_channels_load

)

update all speed for every flow and step guaranteed that channel load ≤ max channel load